requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
openai==1.57.2
httpx==0.28.1
python-multipart==0.0.17
pyyaml==6.0.2
